fix: file references after refactoring
